+++ crates/ruma-client-api/src/session/get_login_types.rs
@@ -12,7 +12,7 @@ pub mod v3 {
 
     use ruma_common::{
         OwnedMxcUri,
-        api::{auth_scheme::NoAuthentication, request, response},
+        api::{auth_scheme::NoAccessToken, request, response},
         metadata,
         serde::{JsonObject, StringEnum},
     };
@@ -24,7 +24,7 @@ pub mod v3 {
     metadata! {
         method: GET,
         rate_limited: true,
-        authentication: NoAuthentication,
+        authentication: NoAccessToken,
         history: {
             1.0 => "/_matrix/client/r0/login",
             1.1 => "/_matrix/client/v3/login",